requirements.txt

streamlit
google-generativeai
PyPDF2
